ci: compliance: split YAMLLint into compliance/ directory
Move YAMLLint out of check_compliance.py and into
scripts/ci/compliance/yaml_lint.py.

The now-unused yamllint import is removed from check_compliance.py.
